[Profiler] Split profiler into a separate module
