+++ Profiler/README.md
@@ -1,0 +1,12 @@
+# Profiler
+
+A small library to measure how long different parts of the 
+code take to run. 
+
+## Usage
+
+A profiling session is defined as the period between a call to 
+`Profiler.beginMeasurements` and a call to `Profiler.endMeasurements`.
+The measured data can then be retrieved using `Profiler.getLastData`.
+You can split the time into subsections using `Profiler.push` to start
+a subsection, and `Profiler.pop` to stop.
