+++ settings.gradle
@@ -4,4 +4,5 @@ include 'Messenger:MessengerClient'
 include 'Messenger:MessengerServer'
 include 'Messenger:MessengerTest'
 include 'MathLib'
+include 'Profiler'
 include 'Robot'
